requirements.txt
annotated-types==0.7.0
anyio==4.8.0
azure-ai-documentintelligence==1.0.0
azure-ai-textanalytics==5.2.0
azure-cognitiveservices-speech==1.41.1
azure-common==1.1.28
azure-core==1.31.0
azure-cosmos==4.9.0
azure-functions==1.21.0
azure-identity==1.18.0
azure-keyvault-secrets==4.9.0
azure-storage-blob==12.23.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
distro==1.9.0
exceptiongroup==1.2.2
grpcio==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
isodate==0.6.1
jiter==0.5.0
lxml==5.3.0
msal==1.31.0
msal-extensions==1.2.0
msrest==0.7.1
oauthlib==3.2.2
openai==1.50.2
portalocker==2.10.1
protobuf==5.29.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMuPDF==1.25.2
python-docx==1.1.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
requests-toolbelt==1.0.0
pandas
openpyxl         # Excel (.xlsx) support for pandas
xlrd             # Older Excel file support (.xls)
PyPDF2           # For reading PDF files
beautifulsoup4   # For parsing HTML
lxml             # Optional but makes BeautifulSoup faster and more robust
azure-storage-blob  # For accessing Azure Blob Storage
beautifulsoup4